New Dispatch Timer Driven UI Updater
Based on an idea from @jimmoyle I've implemented a replacement for the UI Dispatcher which uses secondary variables in the synchash to store the current value of changed properties (kind of how the MVVM pattern would work - abstracting the value of the view to a viewmodel). Current tests show significant improvement - need to validate no other impact.